Update README.md
Advancing seminal vesicle biology by establishing a basic bioniformatic analysis pipeline (DESeq2, PCA)

# Background
The seminal vesicle is the primary contributor to seminal plasma and is responsible for producing an array 
of bioactive constituents which support sperm functioning and influence gene expression in the female reproductive tract. 
Hence, the value of extending the current understanding of the contributions of seminal vesicle secretions to embryo 
health and offspring phenotype cannot be overstated. 

# Contents
The 8 scripts comprising expression_profile_of_mouse_seminal_vesicle analyse the differential expression of
small non-coding RNAs in the mouse seminal vesicle following acrylamide insult, and analyse target gene 
prediction for differentially expressed micro RNAs. 

# Data
The data itself is not included in this repository because I do not have the rights to publish it.